Split team codes into division and team number
Most of the UI was also changed, which results in a large commit.
However, the main reason for this change was to support a "Send to
Division" button on the Notifications page, which was not able to be
included in this commit because it requires a lot of moving around of
the Notifications page.

+++ src/db/subscriptions.py
@@ -10,19 +10,29 @@ from db.models import UserSubscription, db
 # =============================================================================
 
 
-def get_all_teams(email):
+def get_all_subscriptions(email):
     """Gets all the teams the given email is subscribed to."""
     return {
-        (subscription.school, subscription.code): subscription
+        (
+            subscription.school,
+            subscription.division,
+            subscription.number,
+        ): subscription
         for subscription in query(UserSubscription, {"email": email}).all()
     }
 
 
-def get_subscribers(teams):
-    """Gets all the subscribers for the given teams."""
-    teams_subscribers = {team: set() for team in teams}
+def get_all_subscribers(school_team_codes):
+    """Gets all the subscribers for the given school team codes."""
+    teams_subscribers = {
+        school_team_code: set() for school_team_code in school_team_codes
+    }
     for subscription in query(UserSubscription).all():
-        school_team_code = (subscription.school, subscription.code)
+        school_team_code = (
+            subscription.school,
+            subscription.division,
+            subscription.number,
+        )
         if school_team_code not in teams_subscribers:
             continue
         teams_subscribers[school_team_code].add(subscription.email)
@@ -45,56 +55,61 @@ def clear_all_subscriptions():
 # =============================================================================
 
 
-def _get(email, school, team_code):
+def _get(email, school, division, team_number):
     return query(
-        UserSubscription, {"email": email, "school": school, "code": team_code}
+        UserSubscription,
+        {
+            "email": email,
+            "school": school,
+            "division": division,
+            "number": team_number,
+        },
     ).first()
 
 
-def add_team(email, school, team_code):
+def add_team(email, school, division, team_number):
     """Subscribes the given email to the given team.
 
     Returns:
         bool: Whether the operation was successful.
     """
-    existing = _get(email, school, team_code)
+    existing = _get(email, school, division, team_number)
     if existing is not None:
         # already exists; do nothing
         return True
     # add subscription
-    subscription = UserSubscription(email, school, team_code)
+    subscription = UserSubscription(email, school, division, team_number)
     db.session.add(subscription)
     db.session.commit()
     return True
 
 
-def add_school_teams(email, school, team_codes):
+def add_all_teams(email, school_team_codes):
     """Subscribes the given email to all the given teams.
 
     Returns:
         bool: Whether all the operations were successful.
     """
-    existing_teams = get_all_teams(email)
+    existing_subscriptions = get_all_subscriptions(email)
     any_added = False
-    for team_code in team_codes:
-        school_team_code = (school, team_code)
-        if school_team_code in existing_teams:
+    for school_team_code in school_team_codes:
+        if school_team_code in existing_subscriptions:
             continue
         any_added = True
-        subscription = UserSubscription(email, school, team_code)
+        subscription = UserSubscription(email, *school_team_code)
         db.session.add(subscription)
     if any_added:
         db.session.commit()
     return True
 
 
-def remove_team(email, school, team_code):
+def remove_team(email, school, division, team_number):
     """Unsubscribes the given email from the given team.
 
     Returns:
         bool: Whether the operation was successful.
     """
-    existing = _get(email, school, team_code)
+    existing = _get(email, school, division, team_number)
     if existing is None:
         # doesn't exist; do nothing
         return True
@@ -117,6 +132,25 @@ def remove_school_teams(email, school):
         # no subscriptions to remove
         return True
     for subscription in existing_school_teams:
+        db.session.delete(subscription)
+    db.session.commit()
+    return True
+
+
+def remove_division_teams(email, division):
+    """Unsubscribes the given email from all the teams for the given
+    division.
+
+    Returns:
+        bool: Whether all the operations were successful.
+    """
+    existing_division_teams = query(
+        UserSubscription, {"email": email, "division": division}
+    ).all()
+    if len(existing_division_teams) == 0:
+        # no subscriptions to remove
+        return True
+    for subscription in existing_division_teams:
         db.session.delete(subscription)
     db.session.commit()
     return True
